pyproject.toml
[project]
name = "raceline-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "neat-python>=1.0.0",
    "numpy>=2.3.5",
    "scipy>=1.16.3",
]

[tool.ruff]
line-length = 120
